requirements.txt
# astroid==2.15.8
# autopep8==2.0.4
# colorama==0.4.6
# dill==0.3.8
# isort==5.13.2
# lazy-object-proxy==1.10.0
# mccabe==0.7.0
# platformdirs==4.2.0
# pycodestyle==2.11.1
# pylint==2.17.7
python-dateutil==2.8.2
# six==1.16.0
# tomlkit==0.12.3
# wrapt==1.16.0

# Runtime dependencies for the liturgical calendar project
Pillow>=8.0.0
PyYAML>=5.4
requests>=2.25

# On Raspberry Pi, you may need to install system packages for Pillow:
# sudo apt-get install libjpeg-dev zlib1g-dev libfreetype6-dev

# (Dev tools, uncomment if needed)
# pylint==2.17.7
# autopep8==2.0.4
# isort==5.13.2
